examples/local-pipeline/requirements.txt
pipecat-ai[websocket,silero,whisper,piper]==0.0.108
faster-whisper==1.2.1
piper-tts==1.4.2
